fix all regex's ?. to be .?
the ?. was matching the character preceeding and then any one caracter after that. This caused an issue with "(Y|y)a?.(S|s)hou" matching "occupy should" --> because it was saying the a was optional and then there should be one more character (a space). So regex was close, but this difference made this one a little too overaggesive. Fix so now you have to have a ya preceeding the shou in order to match.